Issue-06: Add support for i18n
* Add i18next as dependency
* Move content from `App` to `Home`
* Fix routing
