+++ package.json
@@ -2,7 +2,7 @@
   "name": "gameopsys-react-app",
   "version": "0.1.0",
   "private": true,
-  "homepage": "https://thomasschwengler.github.io/GameOpSysReactApp/",
+  "homepage": "https://tug-cps.github.io/GameOpSysReactApp/",
   "dependencies": {
     "@material-ui/core": "^4.11.4",
     "@material-ui/icons": "^4.11.2",
@@ -16,8 +16,12 @@
     "axios": "^0.21.1",
     "bizcharts": "^4.1.10",
     "history": "^5.0.0",
+    "i18next": "^20.3.2",
+    "i18next-browser-languagedetector": "^6.1.2",
+    "i18next-http-backend": "^1.2.6",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-i18next": "^11.11.0",
     "react-router-dom": "^5.2.0",
     "typescript": "^4.3.4"
   },
