+++ public/locales/de/translation.json
@@ -1,0 +1,5 @@
+{
+  "logout": "Ausloggen",
+  "card_upload_title": "Hochladen",
+  "card_upload_subtitle": "Energieverbrauch hochladen"
+}
