+++ public/locales/en/translation.json
@@ -1,0 +1,5 @@
+{
+  "logout": "Logout",
+  "card_upload_title": "Upload",
+  "card_upload_subtitle": "Upload energy consumption"
+}
