+++ src/Verify.tsx
@@ -4,7 +4,6 @@ import {
     Button,
     Container,
     createStyles,
-    CssBaseline,
     TextField,
     Theme,
     Typography,
@@ -58,7 +57,6 @@ class Login extends React.Component<Props, { password: string }> {
         const {classes} = this.props;
         return (
             <Container component="main" maxWidth="xs">
-                <CssBaseline/>
                 <div className={classes.paper}>
                     <Avatar className={classes.avatar}/>
                     <Typography component="h1" variant="h5">Sign in</Typography>
