+++ src/i18n.tsx
@@ -1,0 +1,22 @@
+import i18n from 'i18next';
+import {initReactI18next} from 'react-i18next';
+import Backend from 'i18next-http-backend';
+import LanguageDetector from 'i18next-browser-languagedetector';
+
+i18n
+    .use(Backend)
+    .use(LanguageDetector)
+    .use(initReactI18next)
+    .init({
+        debug: true,
+        fallbackLng: 'en',
+        interpolation: {
+            escapeValue: false, // not needed for react as it escapes by default
+        },
+        backend: {
+            // for all available options read the backend's repository readme file
+            loadPath: `${process.env.PUBLIC_URL}/locales/{{lng}}/{{ns}}.json`
+        }
+    });
+
+export default i18n;
